Factor response_for_query out of schema.rs
